requirements.txt
PyYAML>=6.0
# TVM 目前版本更新较快，用户可自行选择合适版本；此处仅作示例依赖
# tvm>=0.15.0 